[#19] rename class VariadicResolver -> AssociativeVariadicResolver
casting provided arguments to array can be disabled
